KDA bench: sweep per-sequence length to 16384; cap is HBM, not a realism call
The grid axes are per-sequence length and batch size, and total tokens in a
step is their product -- a batch of long requests legitimately produces a large
product. The previous --max-total-tokens default of 32768 conflated a real HBM
limit with my own assumption that a chunked-prefill step never exceeds 8192
tokens. The second is not a hardware constraint, and applying it cut the sweep
about 25x below what the chip allows.

Re-derived from the working set: ~26KB/token at H=8, K=V=128, BT=64 (8 inputs
+ 2 output + 8 chunk state + ~8 intermediates), so ~20GB of usable HBM is
roughly 800K tokens per step. Default raised to 524288, and seq_lens now reach
16384. Skipped points are still listed.

Also builds the input tensors in float32 rather than numpy's default float64.
They end up bf16, so the float64 staging copy was 4x the final array -- at the
token counts this sweep now reaches that is gigabytes of host memory allocated
and immediately discarded.

+++ benchmark/kernels/kda/utils.py
@@ -57,20 +57,24 @@ def create_kda_uniform_data(
     H, K, V = num_heads, head_dim_k, head_dim_v
 
     def _unit_rows(shape):
-        x = rng.standard_normal(shape)
+        # float32, not the default float64: these tensors end up bf16, and at the
+        # token counts this sweep reaches (N x seq_len, hundreds of thousands) a
+        # float64 staging copy is four times the size of the array being built --
+        # gigabytes of host memory that are immediately thrown away.
+        x = rng.standard_normal(shape, dtype=np.float32)
         return x / np.sqrt((x * x).sum(-1, keepdims=True) + 1e-6)
 
     q = jnp.asarray(_unit_rows((1, total_tokens, H, K)), dtype=dtype)
     k = jnp.asarray(_unit_rows((1, total_tokens, H, K)), dtype=dtype)
-    v = jnp.asarray(rng.standard_normal((1, total_tokens, H, V)), dtype=dtype)
+    v = jnp.asarray(rng.standard_normal((1, total_tokens, H, V), dtype=np.float32), dtype=dtype)
 
-    raw_g = rng.standard_normal((1, total_tokens, H, K))
+    raw_g = rng.standard_normal((1, total_tokens, H, K), dtype=np.float32)
     if lower_bound is not None:
         raw_g = raw_g * gate_stress_scale()
     raw_g = jnp.asarray(raw_g, dtype=dtype)
 
     beta = jax.nn.sigmoid(
-        jnp.asarray(rng.standard_normal((1, total_tokens, H)), dtype=jnp.float32)
+        jnp.asarray(rng.standard_normal((1, total_tokens, H), dtype=np.float32))
     ).astype(dtype)
 
     A_log = jnp.asarray(-1.5 + 0.1 * rng.standard_normal((H,)), dtype=jnp.float32)
